+++ DelvUI/changelog.md
@@ -1,3 +1,6 @@
+# PR
+- Added Buff/Debuffs for the Focus Target
+
 # 0.3.1.2
 - Added option to set Dead Backdrop color for unit- and party frames.